pyproject.toml
[project]
name = "dados-ufpr-2025"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "streamlit",
    "llama-index",
    "google-generativeai",
    "pandas",
    "openpyxl",
    "python-dotenv",
    "plotly>=6.5.0",
    "llama-index-llms-gemini>=0.6.1",
    "llama-index-embeddings-gemini>=0.4.1",
    "tabulate>=0.9.0",
    "llama-index-llms-google-genai>=0.7.4",
    "llama-index-embeddings-google-genai>=0.3.1",
    "nest-asyncio>=1.6.0",
]
